+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 ## v0.0.93 [unreleased]
 
 - set MSRV to 1.95
+- deprecations
+  - new
+    - deprecated `ry.SignedDuration.from_isoformat`,
+      `ry.TimeSpan.parse_common_iso`, and `ry.TimeSpan.from_isoformat` in favor
+      of `*.fromisoformat`; this matches python/pandas
+  - removed
+    - `ryo3_tokio::fs::aiopen` (aka `ry.aiopen`); use `ry.aopen` instead (which
+      has the benefit of not having the letters "ai")
+    - `ryo3_url::URL.replace_*` methods in favor of `ryo3_url::URL.with_*`
+      methods
 
 ---